- Divide the package rsserver to Channel and Item
